Use regexp to remove invalid chars in the filename
This is a simple workaround for #8. It will simply replace characters
considered as invalid with the character _ on the client side. Longer
explaination follows.

Filebin currently uses a client request header to specify the filename.
The filename is limited to ascii characters since request headers are
limited to ascii characters by RFC 822. Proper UTF 8 support will
require a bigger change, possibly including on or both of the following:

1) Transfer the filename using form data instead of a request header.

2) Base64 encode the filename.

It is tempting to use a combination of both, and even to store the files
in the filesystem with base64 encoded names. This will make it possible
to support "folder like" filenames (including /). As mentioned, this is
a bigger change that will require a bit of work and will break the API.