Introduce token based access to reduce abuse
This new feature is disabled by default. By adding --hot-linking=false when starting filebin, token based access is activated. Tokens are generated and distributed when users navigate in filebin, and tokens are valid for a short period of time. This allows users to copy/paste a url and wget it on another computer. On the other hand, sending a direct link to phishing/malware content by email is less interesting with tokens since the token likely has expired when the user opens the link.